+++ src/types/ui/components/index.ts
@@ -1,3 +1,4 @@
 export type * from './accordion.types';
 export type * from './avatar.types';
+export type * from './badge.types';
 export type * from './button.types';