+++ src/ui/components/badge/index.ts
@@ -1,0 +1,3 @@
+import Badge from './badge.component';
+
+export { Badge };